Specifically demand self attribute
