+++ flake8_vedro/visitors/steps_checkers/__init__.py
@@ -8,4 +8,4 @@ from .scope_redefinition_checker import ScopeRedefinitionChecker
 from .then_checker import SingleThenChecker
 from .useless_assert_checker import UselessAssertChecker
 from .when_checker import SingleWhenChecker
-from .mocked_requests_checker import MockedRequestsChecker
+from .mock_assert_checker import MockAssertChecker
